Lab 8 milestone begun
This milestone lab calls for UV's to be done for each and every object, screenshots of the checker on all object in the scene together, and snapshots of the UV wireframe like before

So far I have created an instance version of each  of my objects for the scene, with exception of the walls and the candle flair. I have begun to replace props with instanced versions, so first item on the checklist is to re-instance all the objects into the scene to remove any original or duplicate data. Then next to be done is to UV all the objects. DO NOT FORGET TO RECORD UNWRAPPING THE SIMPLE BOOK.

(I really wish I could be doing this in blender)