Lab 9 and render complete
lab 9 is finished and submitted. Bloom ended up being my hero once again. I am a fan of bloom, it makes lights looks so freaking cool. I'm still glad I used it on squishitron. AHEM anyway, lab 9 is done, so I guess we will be saying goodbye to this simple little room for a while, it was fun while it lasted. But I am honestly excited to progress and further my 3D knowledge and skill.

I also included an extra render with this submission because I thought I found an angle of the candle that looked pretty darn cool.