lab 10 done
Made a plunger for lab 10. Tried a quadsphere for the first time. I think I like it, but need to try the steps in a separate order next time to avoid normals errors. I also tried holding points for the first time so that was fun. Substance is also awesome.

anyway, lab is done and writeup along with it as usual