post 12 stair rail in finishable state
The stair rails are in a state ready to be textured and finished. Allthough a few extra pieces are floating loose that I think might be a couple ways to improve the rails looks near the sloped parts.

at this rate I can only hope the command desk wont be too much of a pain.